requirement.txt
numpy==1.26.4
torch==2.2.1
pandas==2.2.1
pillow==10.2.0
torchvision==0.17.1
tqdm==4.66.2